+++ newton/examples/mpm/example_mpm_snow_ball.py
@@ -388,7 +388,6 @@ class Example:
             "-s",
             type=str,
             default="cg+gauss-seidel",
-            choices=["gauss-seidel", "jacobi", "cg", "cg+jacobi", "cg+gauss-seidel"],
         )
 
         parser.add_argument("--strain-basis", "-sb", type=str, default="P0")